# TestConsoleApp/OpenAPIs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Api服务在线说明"
  },
  "host": "192.168.2.35:8050",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/AgvService/SendTaskToAgv": {
      "post": {
        "tags": [
          "AgvService"
        ],
        "summary": "发送AGV任务服务接口",
        "operationId": "AgvService_SendTaskToAgv",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "agvTasks",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AgvTaskBody"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/AgvService/AgvReport": {
      "post": {
        "tags": [
          "AgvService"
        ],
        "summary": "AGV上报任务状态服务接口",
        "operationId": "AgvService_AgvReport",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AgvReportBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/AgvResponseBody"
            }
          }
        }
      }
    },
    "/api/FactoryWmsService/PrepareProductReq": {
      "post": {
        "tags": [
          "FactoryWmsService"
        ],
        "summary": "厂级Wms成品备货申请",
        "operationId": "FactoryWmsService_PrepareProductReq",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductPrepareRequestBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/FactoryWmsService/ReportProductOutbound": {
      "post": {
        "tags": [
          "FactoryWmsService"
        ],
        "summary": "成品出库信息上报",
        "operationId": "FactoryWmsService_ReportProductOutbound",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReportProductOutboundBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
    "/api/FactoryWmsService/ProductShelvesOrOutReq": {
      "post": {
        "tags": [
          "FactoryWmsService"
        ],
        "summary": "成品上架申请",
        "operationId": "FactoryWmsService_ProductShelvesOrOutReq",
        "consumes": [
          "application/json",
          "text/json"
        ],
        "produces": [
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProductMoveOrOutReqBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WmsResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "AgvTaskBody": {
      "description": "AGV任务信息实体类",
      "type": "object",
      "properties": {
        "Id": {
          "format": "int64",
          "description": "索引；默认与TaskIndex相同",
          "type": "integer"
        },
        "TaskIndex": {
          "format": "int64",
          "description": "任务号",
          "type": "integer"
        },
        "Type": {
          "format": "int32",
          "description": "任务类型；默认104",
          "type": "integer"
        },
        "Priority": {
          "format": "int32",
          "description": "任务优先级；默认0",
          "type": "integer"
        },
        "StartPoint": {
          "description": "起点坐标",
          "type": "string"
        },
        "EndPoint": {
          "description": "终点坐标",
          "type": "string"
        },
        "State": {
          "format": "int32",
          "description": "状态值；默认0",
          "type": "integer"
        }
      }
    },
    "WmsResponseBody": {
      "description": "wms服务接口返回信息",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "请求任务号",
          "type": "string"
        },
        "Code": {
          "description": "状态码",
          "type": "string"
        },
        "Msg": {
          "description": "消息",
          "type": "string"
        }
      }
    },
    "AgvReportBody": {
      "description": "AGV上报信息实体类",
      "type": "object",
      "properties": {
        "TaskIndex": {
          "format": "int64",
          "description": "任务号",
          "type": "integer"
        },
        "State": {
          "format": "int32",
          "description": "状态值；默认0",
          "type": "integer"
        }
      }
    },
    "AgvResponseBody": {
      "description": "AGV服务接口返回信息",
      "type": "object",
      "properties": {
        "Code": {
          "description": "状态码",
          "type": "string"
        },
        "Msg": {
          "description": "消息",
          "type": "string"
        }
      }
    },
    "ProductPrepareRequestBody": {
      "description": "成品备货请求实体类",
      "type": "object",
      "properties": {
        "TaskCode": {
          "description": "申请任务号",
          "type": "string"
        },
        "ReqTime": {
          "description": "申请时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "备货成品信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductPrepareBody"
          }
        }
      }
    },
    "ProductPrepareBody": {
      "description": "备货成品信息",
      "type": "object",
      "properties": {
        "MaterialCode": {
          "description": "物料编码",
          "type": "string"
        },
        "Qty": {
          "description": "数量",
          "type": "string"
        }
      }
    },
    "ReportProductOutboundBody": {
      "description": "成品出库上报实体类",
      "type": "object",
      "properties": {
        "ReqTime": {
          "description": "上报时间",
          "type": "string"
        },
        "MaterialList": {
          "description": "上报信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProductOutboundBody"
          }
        }
      }
    },
    "ProductOutboundBody": {
      "description": "出库成品信息",
      "type": "object",
      "properties": {
        "MaterialBarcode": {
          "description": "物料条码",
          "type": "string"
        },
        "Qty": {
          "format": "double",
          "description": "出库申请：申请数量；返库申请：WMS送料时传入的数量",
          "type": "number"
        }
      }
    },
    "ProductMoveOrOutReqBody": {
      "description": "成品上架或直接出库申请实体类",
      "type": "object",
      "properties": {
        "ReqTime": {
          "description": "上报时间",
          "type": "string"
        },
        "ReqType": {
          "description": "申请类型；Up:上架  Out:直接出库",
          "type": "string"
        },
        "BarcodeList": {
          "description": "条码信息",
          "type": "array",
          "items": {
            "$ref": "#/definitions/BoxBarcodesBody"
          }
        }
      }
    },
    "BoxBarcodesBody": {
      "description": "成品箱条码信息",
      "type": "object",
      "properties": {
        "Barcode": {
          "description": "成品箱条码",
          "type": "string"
        }
      }
    },
...
  },
...
}